refactor: separate form component
